refactor: reorganize Home component structure and move Banner and MiddleSection to components directory
