+++ frontend/src/pages/Home/components/Banner.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styles from "./Banner.module.css";
 
-import heroBanner from "../../../public/assets/images/HeroBanner.png";
+import heroBanner from "../../../../public/assets/images/HeroBanner.png";
 
 function Banner() {
   return (
