+++ frontend/src/pages/Home/components/MiddleSection.module.css
