+++ frontend/src/shared/stationCard/stationDetails/Services.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Services() {
+  return <div></div>;
+}
+
+export default Services;
